+++ passive/ShipTargetStation.cxx
@@ -30,6 +30,7 @@ ShipTargetStation::ShipTargetStation(const char* name,
                                      const Double_t tz,
                                      const TargetVersion tV,
                                      const int nS,
+				     const int HeT,
                                      const char* Title)
     : FairModule(name, Title)
 {
@@ -37,6 +38,7 @@ ShipTargetStation::ShipTargetStation(const char* name,
     fTargetZ = tz;
     fTV = tV;
     fnS = nS;
+    fHeT = HeT;
 }
 
 // -----   Private method InitMedium
@@ -74,14 +76,19 @@ void ShipTargetStation::ConstructGeometry()
 
     InitMedium("H2O");
     TGeoMedium* water = gGeoManager->GetMedium("H2O");
+    InitMedium("Inconel718");
+    TGeoMedium* inc718 = gGeoManager->GetMedium("Inconel718");
 
     InitMedium("vacuums");
     TGeoMedium* vacuum = gGeoManager->GetMedium("vacuums");
 
-    double He_T = 363.15;               // in K, use average 90 degrees C.
+    double He_T = 273.15+fHeT;               // in K, use average 90 degrees C.
     double He_P = 1.6e6 * 6.241509e3;   // 16 bar in MeV/mm3
-    InitMedium("PressurisedHe90");
-    TGeoMedium* pressurised_He = gGeoManager->GetMedium("PressurisedHe90");
+
+    std::ostringstream lHename;
+    lHename << "PressurisedHe" << fHeT;
+    InitMedium(lHename.str().c_str());
+    TGeoMedium* pressurised_He = gGeoManager->GetMedium(lHename.str().c_str());
 
     // CAMM- dirty fix to have pressure and temperature correct for Geant4.
     // Should fix this properly in future...
@@ -91,8 +98,10 @@ void ShipTargetStation::ConstructGeometry()
     TGeoMedium* cooler = (fTV == TargetVersion::Jun25) ? pressurised_He : water;
 
     LOG(INFO) << "-- Target cooler: " << cooler->GetName() << " T=" << cooler->GetMaterial()->GetTemperature()
-              << " K, P=" << cooler->GetMaterial()->GetPressure() << " MeV/mm3";
+              << " K, P=" << cooler->GetMaterial()->GetPressure() << " MeV/mm3, Density="
+	      << cooler->GetMaterial()->GetDensity();
 
+    
     TGeoVolume* tTarget = new TGeoVolumeAssembly("TargetArea");
 
     Double_t zPos = 0.;
@@ -132,6 +141,27 @@ void ShipTargetStation::ConstructGeometry()
         }
     }   // loop on layers
 
+
+    //Target vessel
+    double vessel_thickness = 8 * mm;
+    double vessel_diameter = fDiameter + 150 * mm;
+    double vessel_shift = 76 * mm;
+    double vessel_length = fTargetLength + 2*vessel_shift;
+
+    TGeoVolume* vessel;
+    vessel = gGeoManager->MakeTube("TargetVessel", inc718, vessel_diameter / 2. - vessel_thickness, vessel_diameter / 2., vessel_length / 2.);
+    vessel->SetLineColor(28);
+    tTarget->AddNode(vessel, 1, new TGeoTranslation(0, 0, -1. * vessel_shift + vessel_length / 2.));
+    //Front face
+    vessel = gGeoManager->MakeTube("TargetVesselFront", inc718, 0, vessel_diameter / 2., vessel_thickness / 2.);
+    vessel->SetLineColor(28);
+    tTarget->AddNode(vessel, 1, new TGeoTranslation(0, 0, -1. * vessel_shift - vessel_thickness / 2.));
+    //Back face
+    vessel = gGeoManager->MakeTube("TargetVesselBack", inc718, 0, vessel_diameter / 2., vessel_thickness / 2.);
+    vessel->SetLineColor(28);
+    tTarget->AddNode(vessel, 1, new TGeoTranslation(0, 0, -1. * vessel_shift + vessel_length + vessel_thickness/2.));
+
+    
     // Proximity shielding
 
     double start_of_target = fTargetZ - fTargetLength / 2.;